requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==3.2.18
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pycodestyle==2.10.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
